Logically restructure the mod api
