+++ network-sync/actor_sync.h
@@ -1,0 +1,17 @@
+#ifndef ACTOR_SYNC_H
+#define ACTOR_SYNC_H
+
+#include <stdint.h>
+#include "z64recomp_api.h"
+
+// MARK: - Actor Sync API
+
+void ActorSyncInit();
+const char* ActorSyncGetNetworkId(Actor *actor);
+void ActorSyncRegister(Actor* actor, const char* playerId, int isOwnedLocally);
+
+// MARK: - Internal API (used by callbacks)
+void ActorSyncUpdate(PlayState* play, Actor* actor);
+void ActorSyncProcessRemoteData(PlayState* play);
+
+#endif // ACTOR_SYNC_H
